pyproject.toml
[tool.poetry]
name = "joi-video"
version = "1.0.0"
description = ""
authors = ["Dawson Booth <github@dawsonbooth.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/dawsonbooth/joi-video"
repository = "https://github.com/dawsonbooth/joi-video"
keywords = []
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
include = [
    "LICENSE",
]

packages = [
    { include = "joi_video" },
]

[tool.poetry.scripts]
zip-code-area = 'joi_video.__main__:main'

[tool.poetry.dependencies]
python = "^3.6"
ffmpeg-python = "^0.2.0"
pillow = "^7.2.0"
pymupdf = "^1.17.6"
configparser = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
autopep8 = "^1.5"
pyinstaller = "^3.6"
rope = "^0.17.0"
invoke = "^1.4.1"
pyclean = "^2.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
